+++ relay-config/src/config.rs
@@ -657,6 +657,16 @@ pub struct Limits {
     max_replay_uncompressed_size: ByteSize,
     /// The maximum size for a replay recording Kafka message.
     pub max_replay_message_size: ByteSize,
+    /// The byte size limit up to which Relay will retain
+    /// keys of invalid/removed attributes.
+    ///
+    /// This is only relevant for EAP items (spans, logs, …).
+    /// In principle, we want to record all deletions of attributes,
+    /// but we have to institute some limit to protect our infrastructure
+    /// against excessive metadata sizes.
+    ///
+    /// Defaults to 10KiB.
+    pub max_removed_attribute_key_size: ByteSize,
     /// The maximum number of threads to spawn for CPU and web work, each.
     ///
     /// The total number of threads spawned will roughly be `2 * max_thread_count`. Defaults to
@@ -735,6 +745,7 @@ impl Default for Limits {
             idle_timeout: None,
             max_connections: None,
             tcp_listen_backlog: 1024,
+            max_removed_attribute_key_size: ByteSize::kibibytes(10),
         }
     }
 }
@@ -2467,6 +2478,11 @@ impl Config {
     /// Returns the maximum number of active queries
     pub fn max_concurrent_queries(&self) -> usize {
         self.values.limits.max_concurrent_queries
+    }
+
+    /// Returns the maximum combined size of keys of invalid attributes.
+    pub fn max_removed_attribute_key_size(&self) -> usize {
+        self.values.limits.max_removed_attribute_key_size.as_bytes()
     }
 
     /// The maximum number of seconds a query is allowed to take across retries.