+++ CHANGELOG.md
@@ -10,10 +10,14 @@
 
 **Features**:
 
+- Add new frame fields for MetricKit flamegraphs. ([#5539](https://github.com/getsentry/relay/pull/5539))
 - Apply clock drift correction to logs and trace metrics. ([#5609](https://github.com/getsentry/relay/pull/5609))
+- Add Culture context to event schema. ([#5615](https://github.com/getsentry/relay/pull/5615))
+- Trim spans with a new EAP trimming processor. ([#5616](https://github.com/getsentry/relay/pull/5616))
 
 **Internal**:
 
+- Add EAP double-write for session data. ([#5588](https://github.com/getsentry/relay/pull/5588))
 - Embed AI operation type mappings into Relay. ([#5555](https://github.com/getsentry/relay/pull/5555))
 - Use new processor architecture to process transactions. ([#5379](https://github.com/getsentry/relay/pull/5379))
 - Add `gen_ai_response_time_to_first_token` as a `SpanData` attribute. ([#5575](https://github.com/getsentry/relay/pull/5575))